fix: steam game state

+++ node/native/src/utils.rs
@@ -213,13 +213,13 @@ pub fn from_js<'a>(ctx: &mut FunctionContext<'a>, object: &JsObject) -> io::Resu
         .get(ctx, "total_bytes")
         .ok()
         .and_then(|value| value.downcast::<JsNumber, FunctionContext<'a>>(ctx).ok())
-        .map(|number| number.value(ctx) as i32);
+        .map(|number| number.value(ctx) as i64);
 
     let received_bytes = state
         .get(ctx, "received_bytes")
         .ok()
         .and_then(|value| value.downcast::<JsNumber, FunctionContext<'a>>(ctx).ok())
-        .map(|number| number.value(ctx) as i32);
+        .map(|number| number.value(ctx) as i64);
 
     return Ok(Game {
         _type,
